2dsphere/requirements.txt
PyYAML==3.10
argparse==1.2.1
pymongo==2.4.1
simplejson==3.0.7
web.py==0.37
wsgiref==0.1.2
